+++ NEWS.md
@@ -9,6 +9,7 @@
 - Improved `stocnet_theme()` documentation
 - Improved `match_colors()` documentation
 - Exported `is_dark()` and made it vectorised
+- Added tests for `match_colors()` and `is_dark()`
 
 ## Plotting